requirements.txt
openai>=1.0.0
llama-index>=0.9.0
llama-index-vector-stores-chroma
chromadb>=0.4.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
